Moved auth pages into auth
