+++ MyMoney.Web/ClientApp/src/app/pages/auth/register/register.component.html
