+++ MyMoney.Web/ClientApp/src/app/pages/index.ts
@@ -1,6 +1,6 @@
 export * from './home/home.component';
-export * from './login/login.component';
-export * from './register/register.component';
+export * from './auth/login/login.component';
+export * from './auth/register/register.component';
 export * from './transactions/add/add.transactions.component';
 export * from './transactions/transactions.component';
 export * from './budgets/budgets.component';
